Fix workshop enrollment server-only module split
